Linux Kernel has some interesting tags
The first git tag (created on the linux repo) is weird
because it points to a tree instead of a commit.
It actually pre-dates the commit: https://news.ycombinator.com/item?id=15826636
See it at https://git.kernel.org/pub/scm/linux/kernel/git/torvalds/linux.git/tag/?h=v2.6.11

A bunch of software breaks as a result, including GitHub

All of the below are valid tags on the list of tags:

https://github.com/torvalds/linux/refs-tags/master?source_action=disambiguate&source_controller=files&tag_name=master&q=v2.6.12
https://github.com/torvalds/linux/refs-tags/master?source_action=disambiguate&source_controller=files&tag_name=master&q=v2.6.11

The following either give a 500 or 404:

- https://github.com/torvalds/linux/releases/tag/v2.6.12
- https://github.com/torvalds/linux/releases/tag/v2.6.11
- https://github.com/torvalds/linux/releases/tag/v2.6.11-tree

Jenkins: https://issues.jenkins.io/browse/JENKINS-47350?