Cleanup releases into some standard format.
** Regexes

Regexes help us parse the release into standard forms. A common
regex works for most things, and we can overwrite it on a per
product basis.

Template allows us to use the captured data from the tag to actually
render it into a somewhat standard version name. This means our output
data will not be matching the tag, but should hopefully match whatever
the upstream project uses instead.

We also drop the "v" prefix to stay consistent with how we already
display latest releases on the website