Add a debugging workflow to allow testing
If you push a branch to the main repo, and a branch here with the same
name, then the update script here will use the same branch name instead
of using `main`. This way, we can keep 2 branches in sync and then merge
them.

If a branch does not exist, we'll fallback to cloning master instead