Add GitHub release auto method (closes #34)
This method differs from "git" because it is using the GitHub releases API (https://docs.github.com/en/rest/releases/releases#list-releases) to fetch releases. It gives more accurate releases and release dates because the presence of a git tag does not necessarily means the release is GA.

The following releases are automatically excluded:
- releases that do not match the given regex,
- releases marked as "prereleases" or "draft".

The regex used to extract the version must have exactly one capturing group in order to function properly.