[gke] Adds GKE automation
This avoids using the GKE APIs, much cleaner as a result.
The downside is that we need to keep track of all channels
separately. This also tracks the date a version is made available
on a channel, not the date it is made the default. This is being
slightly optimistic, but I think works well