Fix releasedata.py
Release data were not loaded in the ProductData#__enter__ method. Data would be lost for auto configuration
declaring an auto method updating releases followed by an auto method updating versions.

Also raise an error when product data are completely empty after the update, preventing the product data to be updated at all.
This does not catch all types of errors (what if the second script silently fails completely ?), but that's a start.